src/test/java/org/tmc/data/CuisineJdbcTemplateRepositoryTest.java
package org.tmc.data;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.tmc.models.Cuisine;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
class CuisineJdbcTemplateRepositoryTest {

    @Autowired
    CuisineJdbcTemplateRepository repository;

    @Autowired
    KnownGoodState knownGoodState;

    @BeforeEach
    void setUp() {
        knownGoodState.set();
    }

    @Test
    void shouldFindAll() {
        List<Cuisine> cuisines = repository.findAll();
        assertNotNull(cuisines);

        assertEquals(26, cuisines.size());
    }
}
